Preparing for 1.1.0 release
git-svn-id: https://plugins.svn.wordpress.org/anybackup/trunk@1100840 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/BitsUtil.php
@@ -33,9 +33,11 @@ class BitsUtil {
       exec("renice +$priority ".getmypid());
     }
     static function reset_nice(){
-      // Kill us so that the thread priority will return to the default.  
-      // We cannot raise the priority due to linux standards.
-      posix_kill( getmypid(), 28 );
+      if(function_exists('posix_kill')) {
+        // Kill us so that the thread priority will return to the default.  
+        // We cannot raise the priority due to linux standards.
+        posix_kill( getmypid(), 28 );
+      }
     }
 
     static function all_tables() {
